requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
angelcommunity==2.0.0
    # via cdlib
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
anyio==4.8.0
    # via
    #   httpx
    #   openai
bimlpa==0.1.2
    # via cdlib
cdlib==0.4.0
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.1.1
    # via dynetx
demon==2.0.6
    # via cdlib
distlib==0.3.9
    # via virtualenv
distro==1.9.0
    # via openai
dynetx==0.3.2
    # via cdlib
eva-lcd==0.1.1
    # via cdlib
exceptiongroup==1.2.2
    # via anyio
filelock==3.17.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   virtualenv
fonttools==4.55.8
    # via matplotlib
fsspec==2025.2.0
    # via
    #   huggingface-hub
    #   torch
future==1.0.0
    # via
    #   angelcommunity
    #   dynetx
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
huggingface-hub==0.28.1
    # via
    #   tokenizers
    #   transformers
hydra-core==1.3.2
    # via -r requirements.in
identify==2.6.6
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
igraph==0.11.8
    # via
    #   leidenalg
    #   python-igraph
jinja2==3.1.5
    # via torch
jiter==0.8.2
    # via openai
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.8
    # via matplotlib
leidenalg==0.10.2
    # via -r requirements.in
levenshtein==0.26.1
    # via python-levenshtein
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.0
    # via
    #   -r requirements.in
    #   bimlpa
    #   seaborn
mpmath==1.3.0
    # via sympy
narwhals==1.25.0
    # via plotly
networkx==3.4.2
    # via
    #   -r requirements.in
    #   angelcommunity
    #   bimlpa
    #   cdlib
    #   demon
    #   dynetx
    #   eva-lcd
    #   python-louvain
    #   thresholdclustering
    #   torch
nodeenv==1.9.1
    # via pre-commit
numpy==2.2.2
    # via
    #   -r requirements.in
    #   angelcommunity
    #   cdlib
    #   contourpy
    #   dynetx
    #   eva-lcd
    #   matplotlib
    #   pandas
    #   python-louvain
    #   scikit-learn
    #   scipy
    #   seaborn
    #   thresholdclustering
    #   transformers
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.2
    # via torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
omegaconf==2.3.0
    # via hydra-core
openai==1.61.0
    # via -r requirements.in
packaging==24.2
    # via
    #   huggingface-hub
    #   hydra-core
    #   matplotlib
    #   plotly
    #   pooch
    #   transformers
pandas==2.2.3
    # via
    #   cdlib
    #   seaborn
pillow==11.1.0
    # via matplotlib
platformdirs==4.3.6
    # via
    #   pooch
    #   virtualenv
plotly==6.0.0
    # via cdlib
pooch==1.8.2
    # via cdlib
pre-commit==4.1.0
    # via -r requirements.in
pulp==2.9.0
    # via cdlib
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   openai
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via -r requirements.in
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-igraph==0.11.8
    # via
    #   angelcommunity
    #   cdlib
python-levenshtein==0.26.1
    # via cdlib
python-louvain==0.16
    # via cdlib
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   omegaconf
    #   pre-commit
    #   transformers
rapidfuzz==3.12.1
    # via levenshtein
razdel==0.5.0
    # via -r requirements.in
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   pooch
    #   transformers
safetensors==0.5.2
    # via transformers
scikit-learn==1.6.1
    # via
    #   bimlpa
    #   cdlib
scipy==1.15.1
    # via
    #   cdlib
    #   scikit-learn
seaborn==0.13.2
    # via cdlib
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sympy==1.13.1
    # via torch
texttable==1.7.0
    # via igraph
threadpoolctl==3.5.0
    # via scikit-learn
thresholdclustering==1.1
    # via cdlib
tokenizers==0.21.0
    # via transformers
torch==2.6.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   angelcommunity
    #   cdlib
    #   demon
    #   dynetx
    #   huggingface-hub
    #   openai
    #   transformers
transformers==4.48.2
    # via -r requirements.in
triton==3.2.0
    # via torch
typing-extensions==4.12.2
    # via
    #   anyio
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   torch
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
virtualenv==20.29.1
    # via pre-commit
